Extract views to separate files.
This is beneficial for a few reasons:
- It's easier to find view functions.
- Fewer merge conflicts if multiple people change a view function
  simultaneously
